requirements.txt
Flask    
pandas
matplotlib
seaborn
psycopg2-binary 
pymongo 
python-dotenv